requirements.txt
tornado==6.3.2
mysql-connector-python==8.0.31
openai>=1.0.0
websockets==11.0.3
python-dotenv>=0.19.0